refactoring of cgn merchants api router and payloads
